Remove 'import' and 'export' CLI commands.
The import and export commands are a relic of early Bolt when the file
format was not stable. If the file format changed then users could export
their old data and import it into a new database with a new format.

The Bolt DB file format is stable and will not change so this command is no
longer needed.

Thanks to Alejandro Gaviria for pointing this out.